+++ src/lib/components/chat/Settings/Interface.svelte
@@ -55,7 +55,7 @@
 
 	let collapseCodeBlocks = false;
 	let collapseHistoricalLongResponses = true;
-	let showInlineCitations = true;
+	let showInlineCitations = false;
 	let showMessageOutline = true;
 	let expandDetails = false;
 
@@ -316,7 +316,7 @@
 		collapseCodeBlocks = $settings.collapseCodeBlocks ?? false;
 		collapseHistoricalLongResponses = $settings.collapseHistoricalLongResponses ?? true;
 		responseHtmlFormat = $settings.responseHtmlFormat ?? false;
-		showInlineCitations = $settings.showInlineCitations ?? true;
+		showInlineCitations = $settings.showInlineCitations ?? false;
 		showMessageOutline = $settings.showMessageOutline ?? true;
 		expandDetails = $settings.expandDetails ?? false;